+++ Plugins/UVOCBot.Plugins.Planetside/Objects/CensusQuery/Name.cs
@@ -4,9 +4,9 @@
 /// Initialises a new instance of the <see cref="Name"/> record.
 /// </summary>
 /// <param name="First">The first name.</param>
-/// <param name="Last">The last name.</param>
+/// <param name="FirstLower">A lower-case version of <paramref name="First"/>.</param>
 public record Name
 (
     string First,
-    string Last
+    string FirstLower
 );